requirements.txt
streamlit==1.31.0
mysql-connector-python==8.2.0
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
Pillow==10.2.0